+++ aanalytics2/config.py
@@ -1,12 +1,14 @@
 config_object = {
     "org_id": "",
     "client_id": "",
-    "tech_id": "",
-    "pathToKey": "",
+    "tech_id": None,
+    "pathToKey": None,
     "secret": "",
+    "scopes":None,
     "date_limit": 0,
     "token": "",
-    "tokenEndpoint": "https://ims-na1.adobelogin.com/ims/exchange/jwt"
+    "jwtTokenEndpoint": "https://ims-na1.adobelogin.com/ims/exchange/jwt",
+    "oauthTokenEndpointV2" : "https://ims-na1.adobelogin.com/ims/token/v2"
 }
 orga_admin = {'_org_admin', '_deployment_admin', '_support_admin'}
 header = {"Accept": "application/json",